requirements.txt
datasets
textblob
scikit-learn
matplotlib
streamlit
nltk
rouge
torch
transformers
pillow
ftfy
regex
tqdm
open-clip-torch